Fix stack cradle
Both the stack and the cabal cradle use the same wrapper script which
outputs the working directory as the first argument to fix import
paths. However, the stack cradle was never post processing the
arguments which resulted in the working directory being passed to GHC
which complains that it’s not a module.