Add none cradle
The none cradle is meant to be used to tell your IDE to not even bother
trying to load a certain subdirectory.

For example, if your project contains lots of haskell files for testing
then you don't want to try to load these into your IDE when you open
them. If you're working on a project like GHC, many of them will fail to
load or even parse.