Initialise plugins in frontend hook
The generic frontend already contains this call to do the initialisation
but because we bypassed it, the plugins were never being loaded.

There is probably more we can do to stop repeatedly loading some plugins
but until someone complains, plugins are quite rare so not going to
invest the effort.