add shell and dependency-shell attributes to bios cradle type
This allows passing additional arguments or redirecting a program's
output to collect flags without needing to add a script to the project.
To accomplish that, it utilizes the process package's support for
executing shell commands by passing the CreateProcess data structure
used by it to readProcessWithOutputFile instead of an executable
name and arguments.