Add runGhcLibDir and getRuntimeGhcLibDir
This is part of the work towards getting static binaries properly
working in ghcide and haskell-language-server. To do this we need to be
able to fetch the ghc library directory on the fly, as it will change
with each ghc installation.

getRuntimeGhcLibDir is what client code should use to obtain the libdir,
and searches three places, falling back in this order:
1. the NIX_GHC_LIBDIR environment variable
2. ghc --print-libdir for whatever ghc the cradle is using
3. the libdir baked into ghc-paths

We want to avoid using ghc-paths if possible since it bakes the path
into the binary, which means it isn't portable in the static binary
sense. So if True is passed for the second argument, it will avoid
falling back on ghc-paths, which can be useful when distributing static
binaries etc.

From the library directory we can also get the path to the ghc binary
used, which we can then use to get the cradle ghc version via
getRuntimeGhcVersion