Detect implicit cabal cradle in the absence of cabal.project
A cabal cradle doesn't need a cabal.project file and isn't guaranteed to
have one either. This improves the implicit cradle search for cabal by
falling back to finding .cabal files if hie-bios couldn't find any
cabal.project files.
Because it falls back, if hie-bios tries to query a multi-package cabal
project with a nested cabal package, it will still find use the "larger"
cradle with the upper cabal.project file. See the
implicit-cabal-deep-project test project for an example of this.