Demote process output to Debug severity (#348)
* Demote process output to Debug severity

This is perhaps a more controversial change. The process output is quite
noisy, and arguably not of interest unless you're debugging an issue.

The counterargument is that if something *does* go wrong with running the
cradle command, it's useful to see it. Perhaps we could collect the
output and only emit it on failure?

* Use a special constructor for process output