Keep track of not loaded files for cabal (#453)
When we fail to load some files X, we want to know exactly what files Y we actually did try to load. Where Y ⊆ X.

In the context of HLS, X is the set of files which we already tried to load with the file that we want to additionally load into our HLS session.
Y is a subset of X, chosen based on the Cradle type (bios and stack cradles do not support loading a component with the context information), and GHC and cabal-install version. 
In particular, cabal cradles can make use of the full set of X for initialising a session if and only if the GHC and cabal-install version is recent enough, *and* the user explicitly requested a multi-repl session.

By tracking which files we have actually tried to load, HLS can then try to reduce the number of files we put into sequential loading loop once batch load is failed for haskell/haskell-language-server#4445