Load session using cabal's `--with-repl` command
This avoids quite a lot of hacky implementations to find the correct GHC
version, libdir and ghc-pkg location, as well as global package db
stuff.

There is a corner case which we need to handle:

```
cabal >= 3.15 && lib:Cabal <3.15
```

In other words, when a custom-setup imposes a restriction on the
`lib:Cabal` version, and this `lib:Cabal` version doesn't support the
`--with-repl` flag, then we can't use `--with-repl`.

We detect this by looking at the error message of the cabal command, it
looks roughly like this:

    Error: [Cabal-7107]
    Could not resolve dependencies:
    [__0] trying: cabal-with-custom-setup-0.1.0.0 (user goal)
    [__1] next goal: cabal-with-custom-setup:setup.Cabal (dependency of cabal-with-custom-setup)
    [__1] rejecting: cabal-with-custom-setup:setup.Cabal; 3.10.3.0/installed-3.10.3.0, ... (constraint from --with-repl requires >=3.15)
    ...

We check the error message for the existence of `constraint from
--with-repl requires >=3.15` and then fall back to using the old
approach which doesn't use response files or `--with-repl` flag.

Additionally, we extract a new `Process` module which contains helpers
for cradle actions.